some socketio I will use late to check if the user has left the app or is still connected

+++ app.py
@@ -1,14 +1,18 @@
 from flask import Flask, redirect, url_for, render_template, request, jsonify, make_response
+from flask_socketio import SocketIO
 import os
 import random
 import string
-import logging 
+import logging
+
 
 from db_functions import create_connection, create_creds_and_whitelist_tables
 
 app = Flask(__name__)
-log = logging.getLogger("werkzeug")
-log.disabled = True
+socketio = SocketIO(app)
+
+# log = logging.getLogger("werkzeug")
+# log.disabled = True
 
 
 import sentry_setup_routes, sentry_run_routes
@@ -32,6 +36,7 @@ def entry_page():
 
 
 if __name__ == '__main__':
-	app.run(debug=True,host='0.0.0.0')
-	os.system("rm prev_images/*")
+	socketio.run(app, debug=True,host='0.0.0.0')
+	if len(os.listdir("prev_images")) > 0:
+		os.system("rm prev_images/*")
